+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# 1.4.28
+
+### **Features** 
+
+* `options` property added to `ClsConfig` for applying CSS classes to a single option in a radio groups
+* **brand new sample forms for Material, Kendo, ng-bootstrap and Prime NG**
+* minor template improvements for Material, Kendo, ng-bootstrap and Prime NG
+* initial support for Spinner and InputMask in `ui-primeng`
+* update to ng-bootstrap `beta.4` 
+
+
 # 1.4.27
 
 ### **Bugfixes**